+++ debug_toolbar/static/debug_toolbar/js/toolbar.js
@@ -341,9 +341,9 @@ const djdt = {
                 options.path ? "; path=" + options.path : "",
                 options.domain ? "; domain=" + options.domain : "",
                 options.secure ? "; secure" : "",
-                "sameSite" in options
-                    ? "; sameSite=" + options.samesite
-                    : "; sameSite=Lax",
+                "samesite" in options
+                    ? "; samesite=" + options.samesite
+                    : "; samesite=lax",
             ].join("");
 
             return value;