+++ docs/changes.rst
@@ -7,6 +7,8 @@ Pending
 * Adjusted app directories system check to allow for nested template loaders.
 * Switched from flake8, isort and pyupgrade to `ruff
   <https://beta.ruff.rs/>`__.
+* Converted cookie keys to lowercase. Fixed the ``samesite`` argument to
+  ``djdt.cookie.set``.
 
 4.1.0 (2023-05-15)
 ------------------