+++ docs/changes.rst
@@ -10,7 +10,7 @@ Pending
 * Converted cookie keys to lowercase. Fixed the ``samesite`` argument to
   ``djdt.cookie.set``.
 * Defines the ``BaseStore`` interface for request storage mechanisms.
-* Added the config setting ``TOOLBAR_STORE_CLASS`` to configure the request
+* Added the setting ``TOOLBAR_STORE_CLASS`` to configure the request
   storage mechanism. Defaults to ``debug_toolbar.store.MemoryStore``.
 
 4.1.0 (2023-05-15)